+++ BudgetCalcAPI/Controllers/Transactions/TransactionTypeController.cs
@@ -1,4 +1,7 @@
-﻿using BudgetCalcAPI.Model.Transactions;
+﻿using BudgetCalcAPI.Filters;
+using BudgetCalcAPI.Model.Transactions;
+using BudgetCalcAPI.Service;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -6,6 +9,8 @@ namespace BudgetCalcAPI.Controllers.Transactions
 {
 	[Route("api/transactions/types")]
 	[ApiController]
+	[Authorize]
+	[TokenValidation]
 	public class TransactionTypeController : ControllerBase
 	{
 		private readonly ApplicationContext _context;
@@ -15,11 +20,12 @@ namespace BudgetCalcAPI.Controllers.Transactions
 			_context = context;
 		}
 
-
 		[HttpGet]
 		public async Task<IActionResult> GetTypes()
 		{
-			var types = await _context.TransactionTypes.ToListAsync();
+			var familyGroupId = HttpContext.GetFamilyGroupId();
+
+			var types = await _context.TransactionTypes.Where(t => t.OwnerId == familyGroupId).ToListAsync();
 			return Ok(types);
 		}
 
@@ -27,11 +33,11 @@ namespace BudgetCalcAPI.Controllers.Transactions
 		public async Task<IActionResult> CreateType([FromBody] string typeName)
 		{
 			if (string.IsNullOrEmpty(typeName))
-			{
 				return BadRequest("Type name is null or empty");
-			}
 
-			var newType = new TransactionType { Name = typeName };
+			var familyGroupId = HttpContext.GetFamilyGroupId();
+
+			var newType = new TransactionType { Name = typeName , OwnerId = familyGroupId };
 			_context.TransactionTypes.Add(newType);
 			await _context.SaveChangesAsync();
 			return Ok(newType);
@@ -41,16 +47,14 @@ namespace BudgetCalcAPI.Controllers.Transactions
 		public async Task<IActionResult> DeleteType(int id)
 		{
 			if (id <= 0)
-			{
 				return BadRequest("id is invalid");
-			}
 
 			var type = await _context.TransactionTypes.Where(type => type.Id == id).SingleAsync();
 
 			if (type == null)
-			{
 				return NotFound("Type with given id not found");
-			}
+			if (type.OwnerId != HttpContext.GetFamilyGroupId())
+				return Unauthorized();
 
 			_context.TransactionTypes.Remove(type);
 			await _context.SaveChangesAsync();
